requirements.txt
Django==2.2.5
djangorestframework==3.10.3
flake8==3.7.8
